Refactored ClientLogic into more cohesive classes
