+++ src/main/java/com/rejahtavi/betterflight/BetterFlight.java
@@ -1,7 +1,7 @@
 package com.rejahtavi.betterflight;
 
 import com.rejahtavi.betterflight.client.ClientConfig;
-import com.rejahtavi.betterflight.client.ClientLogic;
+import com.rejahtavi.betterflight.events.ClientEvents;
 import com.rejahtavi.betterflight.client.HUDOverlay;
 import com.rejahtavi.betterflight.common.BetterFlightCommonConfig;
 import com.rejahtavi.betterflight.events.CommonEvents;
@@ -58,7 +58,7 @@ public class BetterFlight {
         eventBus.addListener(this::onLoadCompleteEvent);
 
         if (FMLEnvironment.dist == Dist.CLIENT) {
-            MinecraftForge.EVENT_BUS.register(ClientLogic.class);
+            MinecraftForge.EVENT_BUS.register(ClientEvents.class);
             MinecraftForge.EVENT_BUS.register(HUDOverlay.class);
             eventBus.addListener(this::onClientSetupEvent);
         }
@@ -81,7 +81,7 @@ public class BetterFlight {
 
     @SubscribeEvent
     public void onClientSetupEvent(FMLClientSetupEvent event) {
-        ClientLogic.init();
+        ClientEvents.init();
     }
 
     @SubscribeEvent
