+++ src/main/java/com/rejahtavi/betterflight/network/STCElytraChargePacket.java
@@ -3,8 +3,8 @@ package com.rejahtavi.betterflight.network;
 import java.util.function.Supplier;
 
 import com.rejahtavi.betterflight.BetterFlight;
-import com.rejahtavi.betterflight.client.ClientLogic;
 
+import com.rejahtavi.betterflight.util.ActionHandler;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
@@ -37,7 +37,7 @@ public class STCElytraChargePacket {
     public static void handle(STCElytraChargePacket message, Supplier<NetworkEvent.Context> context) {
         context.get().enqueueWork(() -> {
             DistExecutor.unsafeRunWhenOn(Dist.CLIENT,
-                    () -> () -> ClientLogic.charge = message.getCharge());
+                    () -> () -> ActionHandler.charge = message.getCharge());
         });
         context.get().setPacketHandled(true);
     }
